requirements.txt
fastapi>=0.68.0,<1.0.0
python-multipart>=0.0.5,<1.0.0
minio>=7.1.0,<8.0.0
tika>=1.24,<2.0.0
faiss-cpu>=1.7.0,<2.0.0
numpy>=1.19.0,<2.0.0
elasticsearch>=7.0.0,<8.0.0
sentence-transformers>=2.2.0,<3.0.0
uvicorn>=0.15.0,<1.0.0
pymilvus>=2.0.0,<3.0.0
langchain>=0.0.1,<1.0.0
streamlit>=1.0.0,<2.0.0
python-dotenv>=0.19.0,<1.0.0